Extract FDM solver package
