+++ sandwich_numerical/fdm/solver/__init__.py
@@ -1,0 +1,13 @@
+"""Finite-difference solver primitives."""
+
+from .laplace import set_laplace_update
+from .mesh_block import BoundaryType, MeshBlock
+from .mesh_utils import copy_boundary_gradients, copy_boundary_values
+
+__all__ = [
+    "BoundaryType",
+    "MeshBlock",
+    "copy_boundary_gradients",
+    "copy_boundary_values",
+    "set_laplace_update",
+]
