+++ scripts/run_sandwich_integration.py
@@ -26,7 +26,7 @@ PROJECT_ROOT = Path(__file__).resolve().parents[1]
 if str(PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(PROJECT_ROOT))
 
-from sandwich_numerical.integration import (
+from sandwich_numerical.fdm.integration import (
     GRADIENT_PROFILE_PARABOLIC_ZERO_MEAN,
     GRADIENT_PROFILE_SINE,
     GRADIENT_PROFILES,
